requirements.txt
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
Django==4.2.24
django-cors-headers==4.8.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
PyPDF2==3.0.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
